Add tox-requirements support
Allow the user to provide a path to a requirements file. If set,
tox-package and tox-plugins are ignored and that requirements file
is used as the source of truth for what tox packages to install.
This allows pinning of the tox package and its dependencies.

Document the tox-uv plugin as an approach to speed up dependency
resolution. For now, run-tox itself continues to use pip while the
GitHub Actions to install uv and manage its cache shake out and
stabilize.